Theater Controllers added

+++ src/main/java/com/example/project/bookmyshowbackend/Model/TheaterSeatsEntity.java
@@ -1,14 +1,6 @@
 package com.example.project.bookmyshowbackend.Model;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 import com.example.project.bookmyshowbackend.enums.SeatType;
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -32,7 +24,7 @@ public class TheaterSeatsEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
 
     @Column(name = "seat_number", nullable = false)
     private String seatNumber;
